Add Dijkstra's algorithm
Dijkstra's algorithm comes in two flavours, one that finds the shortest distance between two nodes, and one that finds all distances. Here we have implemented the first flavour, though it would be trivial to add the second if desired.

This algorithm requires knowledge of the whole maze, so does not really fit within the maze/rat convention. As a future task, I shall try modifying Dijkstra's algorithm to cope with incomplete knowledge of the maze, so it can work with maze rats.