Add Tremaux rat
Tremaux needs to know where he is in the maze and which way he is pointing. Then he just marks each end of the passages he goes down. He tries to go down a passage with no marks, or failing that a passage with one mark. He never goes down a passage with two marks.