Multiple rat mazes
The interesting example here is in PairedRat. A single rat that always turns left can get stuck in loops. However, if we introduce two rats with the following rule:

1. If we cannot see the other rat, act like a standard always-left rat
2. If we can see the other rat, go back the way we came (the other rat also reverses)

This seems to solve all mazes, including those with loops. I leave it as an exercise to prove why this is the case.