MemoryRat now remembers everything
Previously, MemoryRat trimmed its memory picture maze to remove dead-ends. It now preserves the whole memory picture, which makes it a lot easier to debug, and stores a set of dead-ends separately. This also simplifies the algorithm.

Also, fix the rendering of the MemoryRat picture so it shows the end correctly, and handles unknown edges.