Improved tests
Sometimes a random rat in a loop goes such a long way round to the exit that the are_mazes_equal comparator blows up due to lack of recursion stack space. Put a test for this in the code. This allows us to add many more tests of equality between mazes